requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.7.0
requests>=2.25.0
PyYAML>=6.0
xgboost>=1.5.0
lightgbm>=3.3.0
matplotlib>=3.4.0
joblib>=1.1.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.62.0 